use sdk cleanly in the indexer
