+++ sdk/packages/sdk/src/intents-helpers.ts
@@ -1,0 +1,5 @@
+// Sub-path entry for tools (e.g. SubQuery indexer) that need intents encoding
+// helpers without the full SDK bundle, which includes TronWeb and crashes in VM2.
+export { decodeERC7821ExecuteBatch, encodeERC7821ExecuteBatch } from "@/protocols/intents/decode-utils"
+export { decodeUserOpScale, encodeUserOpScale } from "@/chains/intentsCoprocessor"
+export { default as IntentGatewayV2 } from "@/abis/IntentGatewayV2"
